Refactor sheet utilities into display modules
